Further cleanup, some renaming and moving refactorings
