+++ dbutils/jmpdb.properties
@@ -1,0 +1,4 @@
+#HSQL Database Engine 2.3.2
+#Sat May 21 17:22:03 CST 2016
+version=2.3.2
+modified=yes
