+++ dbutils/jmpdb.script
@@ -1,0 +1,59 @@
+SET DATABASE UNIQUE NAME HSQLDB54D29B0ACF
+SET DATABASE GC 0
+SET DATABASE DEFAULT RESULT MEMORY ROWS 0
+SET DATABASE EVENT LOG LEVEL 0
+SET DATABASE TRANSACTION CONTROL LOCKS
+SET DATABASE DEFAULT ISOLATION LEVEL READ COMMITTED
+SET DATABASE TRANSACTION ROLLBACK ON CONFLICT TRUE
+SET DATABASE TEXT TABLE DEFAULTS ''
+SET DATABASE SQL NAMES FALSE
+SET DATABASE SQL REFERENCES FALSE
+SET DATABASE SQL SIZE TRUE
+SET DATABASE SQL TYPES FALSE
+SET DATABASE SQL TDC DELETE TRUE
+SET DATABASE SQL TDC UPDATE TRUE
+SET DATABASE SQL TRANSLATE TTI TYPES TRUE
+SET DATABASE SQL CONCAT NULLS TRUE
+SET DATABASE SQL UNIQUE NULLS TRUE
+SET DATABASE SQL CONVERT TRUNCATE TRUE
+SET DATABASE SQL AVG SCALE 0
+SET DATABASE SQL DOUBLE NAN TRUE
+SET FILES WRITE DELAY 500 MILLIS
+SET FILES BACKUP INCREMENT TRUE
+SET FILES CACHE SIZE 10000
+SET FILES CACHE ROWS 50000
+SET FILES SCALE 32
+SET FILES LOB SCALE 32
+SET FILES DEFRAG 0
+SET FILES NIO TRUE
+SET FILES NIO SIZE 256
+SET FILES LOG TRUE
+SET FILES LOG SIZE 50
+CREATE USER SA PASSWORD DIGEST 'd41d8cd98f00b204e9800998ecf8427e'
+ALTER USER SA SET LOCAL TRUE
+CREATE SCHEMA PUBLIC AUTHORIZATION DBA
+SET SCHEMA PUBLIC
+CREATE TEXT TABLE PUBLIC.ACCIDENT_SEVERITY(CODE INTEGER NOT NULL PRIMARY KEY,LABEL VARCHAR(70) NOT NULL)
+CREATE TEXT TABLE PUBLIC.DISTRICT_AUTHORITY(CODE INTEGER NOT NULL PRIMARY KEY,LABEL VARCHAR(70) NOT NULL)
+CREATE TEXT TABLE PUBLIC.LIGHT_CONDITION(CODE INTEGER NOT NULL PRIMARY KEY,LABEL VARCHAR(70) NOT NULL)
+CREATE TEXT TABLE PUBLIC.POLICE_FORCE(CODE INTEGER NOT NULL PRIMARY KEY,LABEL VARCHAR(70) NOT NULL)
+CREATE TEXT TABLE PUBLIC.ROAD_SURFACE(CODE INTEGER NOT NULL PRIMARY KEY,LABEL VARCHAR(70) NOT NULL)
+CREATE TEXT TABLE PUBLIC.WEATHER_CONDITIONS(CODE INTEGER NOT NULL PRIMARY KEY,LABEL VARCHAR(70) NOT NULL)
+CREATE TEXT TABLE PUBLIC.ACCIDENTS(ACCIDENT_INDEX VARCHAR(20) NOT NULL PRIMARY KEY,LONGITUDE NUMERIC(10,6) NOT NULL,LATITUDE NUMERIC(10,6) NOT NULL,POLICE_FORCE INTEGER NOT NULL,ACCIDENT_SEVERITY INTEGER NOT NULL,NUMBER_OF_VEHICLES INTEGER NOT NULL,NUMBER_OF_CASUALTIES INTEGER NOT NULL,DATE DATE NOT NULL,DAY_OF_WEEK INTEGER NOT NULL,TIME VARCHAR(15) NOT NULL,LOCAL_AUTHORITY_DISTRICT INTEGER NOT NULL,LIGHT_CONDITIONS INTEGER NOT NULL,WEATHER_CONDITIONS INTEGER NOT NULL,ROAD_SURFACE_CONDITIONS INTEGER NOT NULL)
+ALTER SEQUENCE SYSTEM_LOBS.LOB_ID RESTART WITH 1
+SET DATABASE DEFAULT INITIAL SCHEMA PUBLIC
+SET TABLE PUBLIC.ACCIDENT_SEVERITY SOURCE 'C:\Users\Alexey\Desktop\EPAMJMP\EPAM-JMP\src\main\resources\accident_severity.csv;ignore_first=true'
+SET TABLE PUBLIC.DISTRICT_AUTHORITY SOURCE 'C:\Users\Alexey\Desktop\EPAMJMP\EPAM-JMP\src\main\resources\district_authority.csv;ignore_first=true'
+SET TABLE PUBLIC.LIGHT_CONDITION SOURCE 'C:\Users\Alexey\Desktop\EPAMJMP\EPAM-JMP\src\main\resources\light_conditions.csv;ignore_first=true'
+SET TABLE PUBLIC.POLICE_FORCE SOURCE 'C:\Users\Alexey\Desktop\EPAMJMP\EPAM-JMP\src\main\resources\police_force.csv;ignore_first=true'
+SET TABLE PUBLIC.ROAD_SURFACE SOURCE 'C:\Users\Alexey\Desktop\EPAMJMP\EPAM-JMP\src\main\resources\road_surface.csv;ignore_first=true'
+SET TABLE PUBLIC.WEATHER_CONDITIONS SOURCE 'C:\Users\Alexey\Desktop\EPAMJMP\EPAM-JMP\src\main\resources\weather_conditions.csv;ignore_first=true'
+SET TABLE PUBLIC.ACCIDENTS SOURCE 'C:\Users\Alexey\Desktop\EPAMJMP\EPAM-JMP\src\main\resources\DfTRoadSafety_Accidents_2009.csv;ignore_first=true'
+GRANT USAGE ON DOMAIN INFORMATION_SCHEMA.SQL_IDENTIFIER TO PUBLIC
+GRANT USAGE ON DOMAIN INFORMATION_SCHEMA.YES_OR_NO TO PUBLIC
+GRANT USAGE ON DOMAIN INFORMATION_SCHEMA.TIME_STAMP TO PUBLIC
+GRANT USAGE ON DOMAIN INFORMATION_SCHEMA.CARDINAL_NUMBER TO PUBLIC
+GRANT USAGE ON DOMAIN INFORMATION_SCHEMA.CHARACTER_DATA TO PUBLIC
+GRANT DBA TO SA
+SET SCHEMA SYSTEM_LOBS
+INSERT INTO BLOCKS VALUES(0,2147483647,0)
