+++ src/main/java/com/epam/extservices/PoliceForceService.java
@@ -1,0 +1,8 @@
+package com.epam.extservices;
+
+/**
+ * Created by Alexey on 25.04.2016.
+ */
+public interface PoliceForceService {
+    String getContactNo(String forceName) throws Exception;
+}
