--- src/main/java/com/epam/service/PoliceForceService.java
@@ -1,8 +1,0 @@
-package com.epam.service;
-
-/**
- * Created by Alexey on 25.04.2016.
- */
-public interface PoliceForceService {
-    String getContactNo(String forceName) throws Exception;
-}
