+++ src/main/resources/config.properties
@@ -11,6 +11,6 @@ processing-thread-num=4
 writing-queue-depth=10
 writing-thread-num=4
 
-#Mock contacts retrieval service
+#Mock contacts retrieval extservices
 max-service-response-time=1
